Revert "refactor(server): extract failed-validation sort into a workspace-migration util"
This reverts commit 0e43c4eab910ff6ab9c8ae96fb4627522721bee5.
